# internal/app/api/core/assets/doc/v0_swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "WireGuard Portal API - UI Endpoints",
        "title": "WireGuard Portal SPA-UI API",
        "contact": {
            "name": "WireGuard Portal Developers",
            "url": "https://github.com/h44z/wg-portal"
        },
        "version": "0.0"
    },
    "basePath": "/api/v0",
    "paths": {
...
        "/interface/peers/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Interface"
                ],
                "summary": "Get peers for the given interface.",
                "operationId": "interfaces_handlePeersGet",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Peer"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
        "/peer/config-mail": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Send peer configuration via email.",
                "operationId": "peers_handleEmailPost",
                "parameters": [
                    {
                        "description": "The peer mail request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.PeerMailRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content if mail sending was successful"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/config-qr/{id}": {
            "get": {
                "produces": [
                    "image/png",
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Get peer configuration as qr code.",
                "operationId": "peers_handleQrCodeGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The peer identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/config/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Get peer configuration as string.",
                "operationId": "peers_handleConfigGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The peer identifier",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/iface/{iface}/all": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Get peers for the given interface.",
                "operationId": "peers_handleAllGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The interface identifier",
                        "name": "iface",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Peer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/iface/{iface}/multiplenew": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Create multiple new peers for the given interface.",
                "operationId": "peers_handleCreateMultiplePost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The interface identifier",
                        "name": "iface",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The peer creation request data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.MultiPeerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.Peer"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/iface/{iface}/new": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Prepare a new peer for the given interface.",
                "operationId": "peers_handleCreatePost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The interface identifier",
                        "name": "iface",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The peer data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Peer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Peer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/iface/{iface}/prepare": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Prepare a new peer for the given interface.",
                "operationId": "peers_handlePrepareGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The interface identifier",
                        "name": "iface",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.Peer"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
        "/peer/iface/{iface}/stats": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Peer"
                ],
                "summary": "Get peer stats for the given interface.",
                "operationId": "peers_handleStatsGet",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The interface identifier",
                        "name": "iface",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/model.PeerStats"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/model.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "model.Error": {
            "type": "object",
            "properties": {
                "Code": {
                    "type": "integer"
                },
                "Message": {
                    "type": "string"
                }
            }
        },
...
        "model.MultiPeerRequest": {
            "type": "object",
            "properties": {
                "Identifiers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "Suffix": {
                    "type": "string"
                }
            }
        },
...
        "model.PeerMailRequest": {
            "type": "object",
            "properties": {
                "Identifiers": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "LinkOnly": {
                    "type": "boolean"
                },
                "PrivateKeys": {
                    "additionalProperties": {
                        "type": "string"
                    },
                    "type": "object"
                }
            }
        },
        "model.PeerStatData": {
            "type": "object",
            "properties": {
                "BytesReceived": {
                    "type": "integer"
                },
                "BytesTransmitted": {
                    "type": "integer"
                },
                "EndpointAddress": {
                    "type": "string"
                },
                "IsConnected": {
                    "type": "boolean"
                },
                "IsPingable": {
                    "type": "boolean"
                },
                "LastHandshake": {
                    "type": "string"
                },
                "LastPing": {
                    "type": "string"
                },
                "LastSessionStart": {
                    "type": "string"
                }
            }
        },
        "model.PeerStats": {
            "type": "object",
            "properties": {
                "Enabled": {
                    "description": "peer stats tracking enabled",
                    "type": "boolean",
                    "example": true
                },
                "Stats": {
                    "description": "stats, map key = Peer identifier",
                    "type": "object",
                    "additionalProperties": {
                        "$ref": "#/definitions/model.PeerStatData"
                    }
                }
            }
        },
...
    }
}